+++ TowerForge/TowerForge/GameViewController.swift
@@ -13,7 +13,7 @@ class GameViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        StorageManager.shared.incrementTotalGamesStarted()
+        AchievementManager.incrementTotalGamesStarted()
         showGameLevelScene(level: 1) // TODO : Change hardcoded level value
     }
